models/requirements.txt
absl-py==2.1.0
aiobotocore==2.12.2
aiohttp==3.9.3
aioitertools==0.11.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
boto3==1.34.86
botocore==1.34.86
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.46.0
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
docker==7.1.0
duckdb==0.9.2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.1
Flask==3.1.0
flatbuffers==23.5.26
fonttools==4.56.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
idna==3.6
importlib_metadata==8.5.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
keras==3.8.0
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==2.20.3
mlflow-skinny==2.20.3
multidict==6.0.5
namex==0.0.8
nest_asyncio==1.5.9
numpy==2.0.2
opencv-python-headless==4.11.0.86
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.14.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.2.0
pluggy==1.4.0
prompt_toolkit==3.0.47
protobuf==5.29.3
psutil==5.9.8
pure_eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
s3fs==2024.3.0
s3transfer==0.10.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack_data==0.6.3
starlette==0.36.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.2
truststore==0.8.0
typing_extensions==4.11.0
tzdata==2025.1
urllib3==2.2.1
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.23.0
